Fixed some nasty bugs that I introduced after changing everything around the last few commits

+++ burst_calc/include/burst_calc/burst_creator.h
@@ -46,7 +46,7 @@ private:
     BurstMerger merger_;
 
     int buffer_size_;
-    int stdev_mult_;
+    double stdev_mult_;
     int burst_window_;
 };
